Give TLD data effective periods, rework generation.
A new `zlint-gtld-update` command is added which fetches ICANN/IANA TLD
data, generating a Go map of TLD names to their effective periods. The
`data/gtld_map.go` file is now generated using `zlint-gtld-update` using
`go generate ./...` in the root of the project. This replaces
`updateTLDs.sh`, `scripts/consolidate_tlds.py`, and the `data/`
directory files.

The `lint_dnsname_right_label_valid_tld.go` lint is updated to check
that a TLD exists and was valid at the time the certificate was issued.
Tests are added for the case where a certificate identifier referenced
a TLD that was not yet valid, that was no longer valid, and that was
valid at the time of issuance.