ci: use GOFLAGS, enforce gofmt -s.
This commit sets up a global `GOFLAGS` to avoid needing to repeat it for
each command.

It also changes the script to exit on the first error, and to enforce
that all non-vendored `.go` files satisfy `gofmt -s` without diffs. This
will make sure all code is consistently formatted and help contributors
do the right thing by default.