tests: add large cert corpus based integration test.
Background
-----------

The idea is to run `zlint` across a big corpus of certificates, ideally
as diverse a corpus as possible, and compare the results against
expected values. The hope is that this will catch regressions in lints
that change lint results for the corpus in an unexpected way.

To start with 60 data files (each of ~5000 certificates) are used to
lint ~600,000 certificates during each CI run to verify lint results do
not change unexpectedly. The tooling was written such that it is easy to
run against larger corpuses before doing a big release, or on a less
frequent schedule than for day to day CI.

The config in this branch adds ~10 minutes runtime to CI. I verified it
would help catch regressions by reverting the fix to the
`e_subject_printable_string_badalpha` lint
(`5dcecad773158b82b5e52064ee2782d1b8a79314`). The integration tests
successfully flagged a difference that would have focused attention on
the new lint: https://gist.github.com/cpu/3a80db08a14a9de7a56db4ff1fc821e9

Golden state
-------------

In a perfect world we would record the full results from each lint for
each certificate and make sure they don't change unexpectedly. Recording
that much data in a "golden" state file in the repo for a large number
of certs would result in pretty significant bloat and slow clones.
Putting it out-of-repo would make PRs cumbersome and adds challenges for
external contributors.

As a compromise the current integration test only tracks a count of
fatal, error, warning, and notice lint results per-certificate but not
details about which lint produced each result at that level.

With the one corpus file in this branch this adds ~70mb of golden state
data to the repo (see `integration/config.json`) which seems acceptable
to me. If we want to trim this down we could stop tracking notice level
lint results or use a more efficient serialized representation than
JSON. One of the advantages of the bloated JSON form is that it is human
readable and easy to update in an editor.

To produce the golden state for the first time or for a large update you
can edit the `integration/config.json` to delete the `"Expected"` map.
When the test is run and there is no configured `Expected` map then the
current results are saved as the `Expected` map to be validated against
in subsequent runs.

Corpus
-------

The cert corpus data is quite large so the integration tests are written
to keep the data out of tree and only download it when needed. Support
is included to automatically uncompress the downloaded data using bzip2
when the URL ends in the `.bz2` file extension.

The Travis CI config is updated to explicitly cache the downloaded
corpus data between builds to help avoid needing to download it every
build. Similarly some effort was made to process data in parallel to
keep builds as fast as possible while still linting as many certs as we
can. Local testing on good hardware can be done even faster by
increasing the parallelism over what is used in CI where the worker
machines are weak and often under-provisioned.

To start with 60 corpus files (each ~5000 certs) are configured for day
to day CI. These files (and more) live in the
https://github.com/zmap/zlint-test-corpus repository.

Running Integration Tests
--------------------------

By default when running `go test` (or `make test`) the integration tests
are not built or run. You must provide `-tag=integration` to build/run
these tests. This should keep day to day development quick.

The `makefile` is updated with a `make integration` target to run just
the integration tests and the Travis script uses that to run integration
tests. The `PARALLELISM` and `INT_FLAGS` makefile variables
can be used to change the number of Go routines and additional flags
(see Extra Command Line Flags) without modifying the Makefile.

Extra Command Line Flags
-------------------

To control the number of linting Go routines used by the `TestCorpus`
integration test change the `-parallelism` command line flag when
running the integration tests. By default 5 Go routines are used for
local integration tests and 3 for Travis CI.

To force corpus data to be downloaded even when it exists on disk use
the `-force` command line flag when running the integration tests. By
default `-force` is disabled.

To have a summary of certificate fingerprint and integration test
results at the end of the integration test provide the `-summary`
command line flag. Note that this is quite spammy and is disabled by
default

To change how many certificates are linted before a '.' character is
printed to the screen use the `-outputTick` flag. By default one period
is printed per 1000 certificates to keep Travis from deciding the CI job
is dead.