deps: update zcrypto, require go 1.13
This updates zcrypto to 7f2fe6f, the tip of master at the time of
writing.

Notably this brings in a change that stores multiple subject CNs and
fingerprint values into the `Certificate` type.

Along the way, bump the Go mod version to 1.13 and update the README to
reflect we expect Go 1.13+ to use zlint master (it's what our CI is
using).