lints: improve template_test.go
Rather than hardcode a mapping of `LintSource` to package directory that
needs to be maintained the `template_test.go` logic should just walk the
filesystem under the `lints/` directory and check all `.go` files.

This makes a smaller `lint` API, removes two places that need to be kept
up to date with new `LintSource`s and results in a test that is robust
against further subdirectory modifications (e.g. a structure deeper than
1 package below `lints`).