project: add goreleaser configuration. (#374)
Adds configuration for GoReleaser to the project/CI.

By default releases are added to the repository in draft status. This gives
maintainers a chance to write the release notes/changelog and verify the build
artifacts before publishing the release. Tags with "rc" in the name will be
marked as pre-release candidates automatically.

Updates #351