util: rewrite test/prepend_testcerts_openssl.sh
The old version needed to be run from a specific directory, didn't pass
`shellcheck`, and would unconditionally prepend the OpenSSL text output
to all certs in the testdata dir (even if they already had it).

This version should be safer and suitable to be integrated with CI in
a later step. It also supports taking a glob as the first argument and
only prepending certs that need it and have a filename that matches
the glob.