util: rewrite test/prepend_testcerts_openssl.sh, update testdata (#421)
* util: rewrite test/prepend_testcerts_openssl.sh

The old version needed to be run from a specific directory, didn't pass
`shellcheck`, and would unconditionally prepend the OpenSSL text output
to all certs in the testdata dir (even if they already had it).

This version should be safer and suitable to be integrated with CI in
a later step. It also supports taking a glob as the first argument and
only prepending certs that need it and have a filename that matches
the glob.

* testdata: add openssl -text output where missed.

This catches up all of the test data to have prepended text.

* testdata: add note to subCertLocalityNameDoesNotNeedToAppear.pem.

This test file does not parse successfully with OpenSSL 1.1.1d on my dev
machine. Adding a small text note about this before the PEM content
avoids the `v2/test/prepend_testcerts_openssl.sh` script emitting
a warning.

* CI: require all testdata is prepended w/ text.

This updates CI to run the `test/prepend_testcerts_openssl.sh` script
and fail if there are any diffs to the `testdata/` directory. This would
indicate there was a `.pem` file that didn't have text prepended to it.