lint: combine ZLint and AWSLabs Sources into Community.
This commit is a breaking API change that combines the `lint.ZLint` and
`lint.AWSLabs` `lint.LintSource`'s into one: `lint.Community`. This
better matches the directory structure we store the lints under and is
more indicative of the fact that we don't intend to have perfect
matching coverage of all awslabs certlint lints.