CI: Add cron workflow for gTLD update PRs.
Previously we used a separate repo with a bash script[0] hooked up to
a bot github user[1] in a Travis CI cron build to automatically create
PRs updating ZLint TLD data on a periodic basis.

Now that we're using Github Actions we can make things much simpler and
self-contained.

This commit adds a `tld-update.yml` workflow that uses
a create-pull-request Github action to replace the separate repo/bash
script/bot user approach.

Not only does this let us delete the bot user's write access to the
ZLint repo but it's also a smarter integration overall and won't
recreate the same PR over and over if it isn't merged right away. Lastly
the Github Actions cron schedule is more flexible so we can run the new
action once every hour instead of just once a day like the Travis
version.

[0]: https://github.com/cpu/zlint-autopull/blob/master/autopull
[1]: https://github.com/tld-update-bot
[2]: https://github.com/peter-evans/create-pull-request