CI: Cleanup hacky tld-update workflow env var use.
The `tld-update.yml` workflow needs the current date in a human readable
form to include in commit messages, PR titles, and the PR body.
Previously this was achieved in a clunky way with an env var echoed into
`$GITHUB_ENV`.

This commit replaces that logic with a cleaner way to achieve the same
thing: setting the date string as the output of a step and then
referencing it later in the workflow.