lints: fix boundary condition in `e_serial_number_longer_than_20_octets` lint (#527)
Previously the `e_serial_number_longer_than_20_octets` lint would mistakenly
pass certificates that had a DER encoded serial number that was exactly 21
octets long. This case typically arises when a serial number is 20 octets long 
with an MSB of 1 since the encoded form will be prefixed with 0x00 to remain
a positive DER encoded integer, thus bumping the encoded length to 21 octets.  
This commit fixes the calculation to correctly return an error finding
for this class of certificates/encoded serial numbers.