lints: return detail for e_ext_duplicate_extension.
Previously the `e_ext_duplicate_extension` lint from the `lint.RFC5280`
source only returned a `lint.Error` result as soon as one duplicate
extension was found in a certificate. It did not indicate which
extension OID was duplicated, or if there was more than one duplicated
extensions.

This commit reworks the lint to do both of these things. The detail
string now indicates all of the extension OIDs that were present more
than once.