Merge from master (#610)
* cmd: add `-version` to `zlint`, `zlint-gtld-update`. (#598)

Our GoReleaser configuration already populates a `version` variable in
LDFLAGS at build time. Prior to this commit we only included the dynamic
version var in the usage output from `--help`. This made it easy to
overlook. We also didn't set the dynamic var from the makefile, leaving
all src builds as the static version "dev".

In this commit we add a `--version` flag to the `zlint` and
`zlint-gtld-update` commands that prints the dynamic version and
exits. We also update the `makefile` so that both binaries get built
with a version that includes the latest tag, and the SHA of the local
git checkout, e.g. `v3.1.0-19-g0807bf95`. This should better match user
expectation for CLI tools.

* lints: fix anyKeyUsage typo in `n_mp_allowed_eku`. (#600)

* deps: update zcrypto to ea3fdbd (#604)

* upgrade to ZCrypto with permissive asn1 parsing

Co-authored-by: Daniel McCarney <daniel@binaryparadox.net>
Co-authored-by: Rob Stradling <rob@sectigo.com>